refactor: move two tests to the reactjs package
because they render a component
