+++ src/reactjs/tests/useQuery.test.tsx
@@ -5675,4 +5675,60 @@ describe('useQuery', () => {
       error,
     })
   })
+
+  it('setQueryData - should not call onSuccess callback of active observers', async () => {
+    const key = queryKey()
+    const onSuccess = jest.fn()
+
+    function Page() {
+      const state = useQuery(key, () => 'data', { onSuccess })
+      return (
+        <div>
+          <div>data: {state.data}</div>
+          <button onClick={() => queryClient.setQueryData(key, 'newData')}>
+            setQueryData
+          </button>
+        </div>
+      )
+    }
+
+    const rendered = renderWithClient(queryClient, <Page />)
+
+    await waitFor(() => rendered.getByText('data: data'))
+    fireEvent.click(rendered.getByRole('button', { name: /setQueryData/i }))
+    await waitFor(() => rendered.getByText('data: newData'))
+
+    expect(onSuccess).toHaveBeenCalledTimes(1)
+    expect(onSuccess).toHaveBeenCalledWith('data')
+  })
+
+  it('setQueryData - should respect updatedAt', async () => {
+    const key = queryKey()
+
+    function Page() {
+      const state = useQuery(key, () => 'data')
+      return (
+        <div>
+          <div>data: {state.data}</div>
+          <div>dataUpdatedAt: {state.dataUpdatedAt}</div>
+          <button
+            onClick={() =>
+              queryClient.setQueryData(key, 'newData', { updatedAt: 100 })
+            }
+          >
+            setQueryData
+          </button>
+        </div>
+      )
+    }
+
+    const rendered = renderWithClient(queryClient, <Page />)
+
+    await waitFor(() => rendered.getByText('data: data'))
+    fireEvent.click(rendered.getByRole('button', { name: /setQueryData/i }))
+    await waitFor(() => rendered.getByText('data: newData'))
+    await waitFor(() => {
+      expect(rendered.getByText('dataUpdatedAt: 100')).toBeInTheDocument()
+    })
+  })
 })
